CLIP/requirements.txt
fastapi
uvicorn
torch
torchvision
ftfy
regex
tqdm
Pillow
git+https://github.com/openai/CLIP.git
python-multipart